% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{allpdiff}
\alias{allpdiff}
\title{All pairwise differences}
\usage{
allpdiff(x, y, paired = FALSE, na.rm = TRUE)
}
\arguments{
\item{x, y}{Two vectors of same or different lengths.}

\item{paired}{If paired = TRUE and na.rm = TRUE, pairs are removed.}

\item{na.rm}{Missing values are removed if na.rm = TRUE.}
}
\description{
Calculate all pairwise differences between 2 vectors.
}
\section{Note}{

Combine functions \code{loc2dif} & \code{wmwloc} from
Rallfun-v31.txt - see \url{https://github.com/nicebread/WRS}.
}

