% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{annotate_quartiles}
\alias{annotate_quartiles}
\title{Add text annotation of quartiles}
\usage{
annotate_quartiles(p, data, x = 0, hjust = 0, vjust = 0, size = 10)
}
\description{
Add text annotation of quartiles
}

