% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cid}
\alias{cid}
\title{Cliff's delta test}
\usage{
cid(x, y, alpha = 0.05)
}
\arguments{
\item{x, y}{Two vectors. Missing values are automatically removed.}

\item{alpha}{Alpha significance level.}
}
\description{
Compute Cliff's analog of WMW test. Compute a confidence interval for delta
using the method in Cliff, 1996, p. 140, eq 5.12.
The null hypothesis is that for two independent group, P(X<Y)=P(X>Y).
This function reports a 1-alpha confidence interval for
P(X>Y)-P(X<Y).
}
\details{
\code{cid} returns a list containing the value of the test statistic, its
confidence interval, and its p value.
}
\section{Note}{

From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}
}

\references{
Cliff, N. (1996) Ordinal methods for behavioral data analysis. Erlbaum, Mahwah, N.J.
}
\seealso{
To compute a p value, use \code{cidv2}.
}
