% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cidv2}
\alias{cidv2}
\title{P value of Cliff's delta test}
\usage{
cidv2(x, y, alpha = 0.05)
}
\arguments{
\item{x, y}{Two vectors. Missing values are automatically removed.}

\item{alpha}{Alpha significance level.}
}
\description{
Compute Cliff's analog of WMW test & associated p value
}
\details{
\code{cidv2} returns a list containing the value of the test statistic, its
confidence interval, and its p value.
}
\section{Note}{

From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}
}
\references{
Cliff, N. (1996) Ordinal methods for behavioral data analysis. Erlbaum, Mahwah, N.J.
Wilcox, R.R. (2011) Inferences about a Probabilistic Measure of Effect Size When Dealing with More Than Two Groups. Journal of Data Science, 9, 471-486.
}
\seealso{
This function uses the function cid.
}

