% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ecdf}
\alias{ecdf}
\title{Compute empirical cdf
 Compute the empirical cdf for data in x evaluated at val.
 That is, estimate P(X <= val)}
\usage{
ecdf(x, val)
}
\description{
Compute empirical cdf
 Compute the empirical cdf for data in x evaluated at val.
 That is, estimate P(X <= val)
}

