% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flp}
\alias{flp}
\title{Reaction time data from the French Lexicon Project}
\format{
A data frame with 1917983 rows and 4 variables:
\describe{
  \item{participant}{id factor, from 1 to 959}
  \item{rt}{single trial reaction times, in ms}
  \item{acc}{accuracy, 1 for correct and 0 for incorrect responses}
  \item{condition}{condition factor, with levels "word" and "non-word"}
}
}
\source{
\url{https://sites.google.com/site/frenchlexicon/results}
}
\usage{
flp
}
\description{
A dataset containing reaction times data from 959 participants in 2 conditions
from a lexical decision task. File created using the script getflprtdata.Rmd
available in /data-raw/.
}
\keyword{datasets}
