% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hd}
\alias{hd}
\title{Harrell-Davis quantile estimator}
\usage{
hd(x, q = 0.5, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of continuous observations.}

\item{q}{A quantile between 0 and 1 - default is 0.5, the median.}

\item{na.rm}{Set to TRUE to remove NA, FALSE otherwise - default is TRUE.}
}
\description{
Compute the Harrell-Davis estimate of the qth quantile.
}
\details{
\code{hd} is a weighted average of all the order statistics. See
\url{https://garstats.wordpress.com/2016/06/09/the-harrell-davis-quantile-estimator/}
for more details.
}
\examples{
hd(x) # use default estimation of the median (q = 0.5)
hd(x, q = 0.25) # estimate the first quartile
hd(x, q = 0.75) # estimate the third quartile
}
\references{
Harrell, F.E. & Davis, C.E. (1982) A new distribution-free quantile estimator. Biometrika, 69, 635-640.
}

