% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hdpbci}
\alias{hdpbci}
\title{Bootstrap confidence interval of the qth quantile}
\usage{
hdpbci(x, q = 0.5, alpha = 0.05, nboot = 2000, nv = 0)
}
\arguments{
\item{x}{A vector of continuous observations.}

\item{q}{A quantile between 0 and 1 - default = 0.5, meaning a confidence
interval for the median is computed.}

\item{alpha}{Alpha level - default = 0.05, such that a 95% confidence
interval is computed.}

\item{nboot}{Number of bootstrap samples - default = 2000.}

\item{nv}{Null value when computing a p-value.}
}
\value{
A list with variables: \code{q} \code{estimate}, \code{ci}, \code{n},
  \code{p.value}
}
\description{
Compute a percentile bootstrap confidence interval for the qth quantile via
the Harrell--Davis estimator. Appears to be best method when there are tied
values.
}
\section{Note}{
 Adaptation of \code{qcipb} from Rallfun-v33.txt. See
  \url{https://github.com/nicebread/WRS/}
}

\seealso{
\code{\link{hd}}
}
