% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hdseq}
\alias{hdseq}
\title{Compute quantiles}
\usage{
hdseq(x, qseq = seq(0.1, 0.9, 0.1))
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{q}{A numeric vector of quantiles - default to the deciles.}
}
\value{
A vector of quantile estimates.
}
\description{
Estimate the quantiles for the data in vector x using the Harrell-Davis
estimate of the qth quantile.
}
\examples{
x <- rnorm(100) # create vector
hdseq(x) # estimates of the deciles
hdseq(x, qseq = c(0.25, 0.5, 0.75)) # estimates of the quartiles
hdseq(x, qseq = c(seq(1,4),seq(6,9))) # estimates of the deciles except the median

}
