% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsf_calc.R
\name{hsf_pb}
\alias{hsf_pb}
\title{Percentile bootstrap hierarchical shift function for one group, two dependent conditions}
\usage{
hsf_pb(
  data = df,
  formula = obs ~ cond + id,
  qseq = seq(0.1, 0.9, 0.1),
  tr = 0.2,
  alpha = 0.05,
  qtype = 8,
  todo = c(1, 2),
  nboot = 1000
)
}
\arguments{
\item{data}{A data frame in long format. Missing values are not allowed.}

\item{formula}{A formula with format response variable ∼ predictor variable + id,
where ~ (tilde) means "is modeled as a function of" and '+ id' indicates the variable containing the participants' id number.}

\item{qseq}{Quantiles to estimate - default = deciles.}

\item{tr}{Percentage of trimming, value between 0 and 1 - default = 0.2 = 20\%. Set to zero to get results for the mean.}

\item{alpha}{Alpha level - default 0.05.}

\item{qtype}{Type of quantile estimation algorithm to pass to \code{quantile} - default = 8.}

\item{todo}{Order of the groups to compare - default = 1 minus 2.}

\item{nboot}{Number of boostrap samples - default = 1000.}
}
\value{
A list of 8 results:
\enumerate{
  \item \strong{comparison}: names of two conditions being compared.
  \item \strong{individual_sf}: shift functions for every participant.
  \item \strong{group_differences}: group quantile differences.
  \item \strong{ci}: group confidence intervals for quantile differences.
  \item \strong{hdi}: group highest density intervals for quantile differences.
  \item \strong{quantiles}: quantiles estimated in each participant and condition.
  \item \strong{boot_samples}: bootstrap differences for each quantile.
  \item \strong{nboot}: number of bootstrap samples.
  }
}
\description{
\code{hsf_pb} returns a percentile bootstrap hierarchical shift function for
one group of participants, tested in two dependent conditions (see
\href{https://github.com/GRousselet/rogme/blob/master/docs/hsf.md}{vignette}
on github). Full distributions of measurements must be available for each
participant and condition. First, quantiles are computed for the distribution
of measurements from each condition and each participant. Second, the
quantiles are subtracted in each participant. Third, a trimmed mean is
computed across participants for each quantile. Confidence intervals are
computed using the percentile bootstrap. This function focuses on estimation, so
there is no correction for multiple comparisons and no p values. Plot the
shift function using \code{plot_hsf_pb} or \code{plot_hsf_pb_dist}.
}
\examples{
set.seed(22) # subset random sample of participants from the French Lexicon Project
id <- unique(flp$participant)
df <- subset(flp, flp$participant \%in\% sample(id, 50, replace = FALSE))
out <- hsf_pb(df, rt ~ condition + participant) # use the default parameters
plot_hsf_pb(out) # plot results. Shift functions are overall negative, as
participants tend to be faster in the Word condition than in the Non-Word
condition.

out <- hsf_pb(df, rt ~ condition + participant, qseq = c(.25, .5, .75)) # estimate quartiles only

out <- hsf_pb(df, rt ~ condition + participant, todo = c(2,1)) # reverse comparison

}
\references{
Rousselet, G. A., & Wilcox, R. R. (2019, January 17). Reaction
  times and other skewed distributions: problems with the mean and the
  median. https://doi.org/10.31234/osf.io/3y54r
}
\seealso{
\code{\link{plot_hsf_pb}} to plot the results.
\code{\link{plot_hsf_pb_dist}} to plot the distributions of bootstrap samples.
}
