% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ks}
\alias{ks}
\title{Compute the Kolmogorov-Smirnov test statistic}
\usage{
ks(x, y, w = FALSE, sig = TRUE, alpha = 0.05)
}
\arguments{
\item{x, y}{Two vectors. Missing values are automatically removed.}

\item{w}{TRUE or FALSE. Set w to FALSE to use the standard version. Set w to
TRUE to use a weighted version, which is more sensitive to differences
occuring in the tails of the distributions.}

\item{sig}{TRUE to use the exact significance level. If there are ties, the
reported significance level is exact when using the unweighted test, but
for the weighted test the reported level is too high.}
}
\description{
\code{ks} returns a list containing the value of the test statistic, the
approximate .05 critical value, and the exact significance level if sig=T.
}
\section{Note}{

From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}
}

\seealso{
This function uses the functions ecdf, kstiesig, kssig and kswsig
}
