% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kssig}
\alias{kssig}
\title{Compute significance level of the  Kolmogorov-Smirnov test statistic
m=sample size of first group
n=sample size of second group
val=observed value of test statistic}
\usage{
kssig(m, n, val)
}
\description{
Compute significance level of the  Kolmogorov-Smirnov test statistic
m=sample size of first group
n=sample size of second group
val=observed value of test statistic
}
