% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kstiesig}
\alias{kstiesig}
\title{Compute significance level of the Kolmogorov-Smirnov test statistic}
\usage{
kstiesig(x, y, val)
}
\description{
Compute significance level of the Kolmogorov-Smirnov test statistic
for the data in vectors x and y at val, the observed value of the
test statistic.
\code{kstiesig} allows ties among values.
}
