% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kswsig}
\alias{kswsig}
\title{Compute significance level of the weighted
Kolmogorov-Smirnov test statistic}
\usage{
kswsig(m, n, val)
}
\arguments{
\item{m}{Sample size of first group}

\item{n}{Sample size of second group}

\item{val}{Observed value of test statistic}
}
\description{
Compute significance level of the weighted
Kolmogorov-Smirnov test statistic
}
