% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkt1}
\alias{mkt1}
\title{Make tibble (data frame) for one group}
\usage{
mkt1(x, name = c("data", "gr"))
}
\arguments{
\item{x}{A vector.}

\item{gr_names}{The name of the first column - default "gr" for groups.}

\item{obs_names}{The name of the second column - default "obs" for
observations.}
}
\value{
A tibble - see details in \code{\link[TIBBLE]{tibble}}.
}
\description{
Make tibble from one vector x. The vector is re-organised in a
data frame with column 1 for group label and column 2 for observations.
}
\examples{
# Generate one vector
g1 <- rnorm(10)

# Make tibble using default parameters
df <- mkt1(g1)

# Make tibble using custom parameters
df <- mkt1(g1, gr_names = "groups", obs_names = "values", group_label =
"group1")

# For instance, vector x:
x <- c(1, 2, 3, 4, 5)
df <- mkt1(x)
# become a tibble: 5 × 2:
       gr   obs
   <fctr> <dbl>
1  Group1     1
2  Group1     2
3  Group1     3
4  Group1     4
5  Group1     5
}

