% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkt2}
\alias{mkt2}
\title{Make tibble (data frame) for two groups}
\usage{
mkt2(x, y, gr_names = "gr", obs_names = "obs", group_labels = c("Group1",
  "Group2"))
}
\arguments{
\item{x, y}{Two vectors.}

\item{gr_names}{The name of the first column - default "gr" for groups.}

\item{obs_names}{The name of the second column - default "obs" for
observations.}
}
\value{
A tibble - see details in \code{\link[TIBBLE]{tibble}}.
}
\description{
Make tibble from two vectors x and y. The two vectors are re-organised in a
data frame with column 1 for group labels as factors and column 2 for observations.
}
\examples{
# Generate two vectors
g1 <- rnorm(10)
g2 <- rnorm(10)

# Make tibble using default parameters
df <- mkt2(g1, g2)

# Make tibble using custom parameters
df <- mkt2(g1, g2, gr_names = "groups", obs_names = "values", group_labels =
c("group1", "group2"))

# For instance, vectors x & y:
x <- c(1, 2, 3, 4, 5)
y <- c(11, 12, 13, 14, 15)
df <- mkt2(x, y)
# become a tibble: 10 × 2:
       gr   obs
   <fctr> <dbl>
1  Group1     1
2  Group1     2
3  Group1     3
4  Group1     4
5  Group1     5
6  Group2    11
7  Group2    12
8  Group2    13
9  Group2    14
10 Group2    15
}
