% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{onsets}
\alias{onsets}
\title{ERP onsets from 120 participants}
\format{A vector of 120 values in ms.}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/ejn.13100}
}
\usage{
onsets
}
\description{
A dataset containing estimated onsets of face - noise ERP differences
from 120 participants
}
\keyword{datasets}
