% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{onsets2}
\alias{onsets2}
\title{ERP onsets from 120 participants, 74 tested twice.}
\format{A matrix of 120 rows of onset values in ms and 2 columns for sessions 1 and 2. Missing values in session 2 are marked as NA.}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/ejn.13100}
}
\usage{
onsets2
}
\description{
A dataset containing estimated onsets of face - noise ERP differences
from 120 participants in session 1 and 74 participants in session 2.
}
\keyword{datasets}
