% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pb2gen}
\alias{pb2gen}
\title{Compare two independent groups using percentile bootstrap}
\usage{
pb2gen(x, y, alpha = 0.05, nboot = 2000, est = hd, ...)
}
\arguments{
\item{x, y}{Two vectors of continuous variables.}

\item{alpha}{Alpha level - default = 0.05, which means that a 95% confidence
interval is computed.}

\item{nboot}{Number of bootstrap samples - default = 2000.}
}
\value{
A list with variables: \code{est.1} \code{est.2}, \code{est.diff} \code{ci}, \code{p.value}, \code{sq.se}, \code{n1}, \code{n2}
}
\description{
Compute a bootstrap confidence interval for the
the difference between any two parameters corresponding to
independent groups.
By default, the Harrell-Davis estimates of the median are compared.
}
\section{Note}{
 Adaptation of \code{pb2gen} from Rallfun-v33.txt. See
  \url{https://github.com/nicebread/WRS/}
}
\seealso{
\code{\link{hd}}
}

