% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pdata}
\alias{pdata}
\title{Fake data from 35 participants}
\format{An object of class \code{data.frame} with 35 rows and 3 columns.}
\usage{
pdata
}
\description{
A dataset containing fake data from 35 participants in 2 conditions.
}
\details{
@format A data frame with 35 rows and 3 variables:
 \describe{
   \item{participant}{id factor, from 1 to 35}
   \item{condition1}{results from condition 1, in arbitrary units}
   \item{condition2}{results from condition 2, in arbitrary units}
 }

@source \url{https://www.biorxiv.org/content/10.1101/121079v2}
}
\keyword{datasets}
