% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_dec_bars}
\alias{plot_dec_bars}
\title{Add bars marking the deciles to plot}
\usage{
plot_dec_bars(p, col = "grey21", width = 0.5, dec_size = 0.5,
  md_size = 1, alpha = 1)
}
\description{
Add bars marking the deciles to plot
}

