% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_dec_ci}
\alias{plot_dec_ci}
\title{Plot deciles and their confidence intervals
using the output of `quantiles_pbci`
GAR, University of Glasgow, 2016-07-15}
\usage{
plot_dec_ci(out = out, plotzero = TRUE, xtitle = "Differences",
  hjust = -0.05, vjust = 0.2, size = 6)
}
\description{
Plot deciles and their confidence intervals
using the output of `quantiles_pbci`
GAR, University of Glasgow, 2016-07-15
}

