% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_dec_links}
\alias{plot_dec_links}
\title{Add deciles to plot + links between deciles}
\usage{
plot_dec_links(p, sf = sf, dec_col = "grey21", dec_width = 0.5,
  dec_size = 0.5, md_size = 1, link_col = c("darkviolet", "darkorange2"),
  link_alpha = c(0.4, 1), add_rect = FALSE, rect_alpha = NULL,
  rect_col = NULL, add_lab = FALSE, labres = 2)
}
\description{
Used in README to demonstrate the shift function.
}
