% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_asymmetry_plot.R
\name{plot_diff_asym}
\alias{plot_diff_asym}
\title{Plot difference asymmetry function}
\usage{
plot_diff_asym(data = df, xlabel = "Quantiles",
  ylabel = "Quantile sum = q + 1-q", symb_col = NULL, symb_fill = NULL,
  symb_size = 5, symb_shape = 21, diffint_col = NULL,
  diffint_size = 0.5, dec_line_col = NULL, dec_line_alpha = 0.5,
  dec_line_size = 1.5)
}
\description{
Plot difference asymmetry function using output from
Rand Wilcox's qwmwhd or difQpci functions. For naming consistency,
these functions have been renamed `asymhd` and `asymdhd` in `rogme`.
}
