% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_hd_bars}
\alias{plot_hd_bars}
\title{Add bars marking quantiles to plot
Default to deciles.
To plot the deciles with a different bar size for the median,
use \code{\link{plot_dec_bars}}.}
\usage{
plot_hd_bars(p, q_seq = seq(0.1, 0.9, 0.1), col = "grey21", width = 0.5,
  size = 0.5)
}
\description{
Add bars marking quantiles to plot
Default to deciles.
To plot the deciles with a different bar size for the median,
use \code{\link{plot_dec_bars}}.
}

