% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_hd_links}
\alias{plot_hd_links}
\title{Add bars marking quantiles + links between quantiles to ggplot object created
by \code{\link{plot_scat2}}. Used in README to demonstrate the shift
function.}
\usage{
plot_hd_links(p, sf = sf, q_col = "grey21", q_width = 0.5,
  q_size = 0.5, md_size = 1, link_col = c("darkviolet",
  "darkorange2"), link_alpha = c(0.4, 1), add_rect = FALSE,
  rect_alpha = NULL, rect_col = NULL, add_lab = FALSE, labres = 2,
  text_size = 5)
}
\arguments{
\item{p}{A ggplot object returned by \code{\link{plot_scat2}}.
Used in README to demonstrate the shift function.}

\item{sf}{A shift function for the 2 groups illustrated in p.}

\item{q_col}{Colour of the bars}

\item{q_width}{Length of the bars}

\item{q_size}{Thickness of the bars, except for the median}

\item{md_size}{Thickness of the median bar}

\item{link_col}{Colour of the links between quantiles - default "darkviolet"
for negative differences, "darkorange2" for postivie differences}

\item{link_alpha}{Alpha transparency of the links between quantiles - default
c(0.4, 1), maximum value for the median, decreasing towards the lowest
value for the extreme quantiles.}

\item{add_rect}{Add rectangle marking the location of the quantiles in group 1.}

\item{rect_alpha}{Alpha transparency for the rectangle.}

\item{rect_col}{Colour for the rectangle}

\item{add_lab}{If TRUE, add labels for differences between extreme quantiles - default = FALSE}

\item{labres}{Number of decimales for the labels - default = 2}

\item{text_size}{Size of the labels - default = 5}
}
\description{
Add bars marking quantiles + links between quantiles to ggplot object created
by \code{\link{plot_scat2}}. Used in README to demonstrate the shift
function.
}
