% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsf_plot.R
\name{plot_hsf_pb}
\alias{plot_hsf_pb}
\title{Plot percentile bootstrap hierarchical shift function}
\usage{
plot_hsf_pb(data, null_value = 0, interv = "hdi",
  viridis_option = "B", ind_line_alpha = 0.5, ind_line_size = 0.5,
  gp_line_colour = "black", gp_line_size = 1,
  gp_point_colour = "black", gp_point_size = 0.5)
}
\arguments{
\item{data}{A list generated by \code{\link{hsf_pb}}.}

\item{null_value}{Reference value to plot as a horizontal line - default = 0.}

\item{interv}{Interval to plot: "ci" or "hdi" - default = "hdi"}

\item{viridis_option}{Viridis colour map for individual shift functions - default = "B".}

\item{ind_line_alpha}{Alpha transparency for individual shift functions - default = 0.5}

\item{ind_line_size}{Line thickness for for individual shift functions - default = 0.5}

\item{gp_line_colour}{Group line colour - default = "black".}

\item{gp_line_size}{Group line size - default = 1.}

\item{gp_point_colour}{Group point colour - default = "black".}

\item{gp_point_size}{Group point size - default = 0.5}
}
\value{
A ggplot object.
}
\description{
Plot percentile bootstrap hierarchical shift function generated with \code{\link{hsf_pb}}.
The function returns a ggplot object, which can be
customised using the \href{http://docs.ggplot2.org/current/}{ggplot2}
package. Individual shift functions are superimposed in the background. The trimmed mean of the shift functions is plotted on top, with a percentile bootstrap confidence interval or highest density interval for each quantile.
A horizontal line indicates a null reference value.
}
\examples{
plot_hsf_pb(out) # default plot
p <- plot_hsf_pb(out) # get ggplot object
plot_hsf_pb(out, gp_line_colour = "purple", gp_point_colour = "purple") # pink group shift function

}
\references{
Rousselet, G. A., & Wilcox, R. R. (2019, January 17). Reaction
  times and other skewed distributions: problems with the mean and the
  median. https://doi.org/10.31234/osf.io/3y54r
}
