% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsf_plot.R
\name{plot_hsf_pb_dist}
\alias{plot_hsf_pb_dist}
\title{Plot hierarchical shift function percentile bootstrap samples}
\usage{
plot_hsf_pb_dist(data, null_value = 0, point_interv = "mode_hdi",
  interval_width = c(0.5, 0.9), fill_colour = "orange",
  int_colour = "black")
}
\arguments{
\item{data}{A list generated by \code{\link{hsf_pb}}.}

\item{null_value}{Reference value to plot as a horizontal line - default = 0.}

\item{point_interv}{Option to pass to \href{https://mjskay.github.io/tidybayes/reference/point_interval.html}{point_interval} - default = "mode_hdi"}

\item{interval_width}{Width of the intervals - default c(0.5, 0.9)}

\item{fill_colour}{Colour of the bootstrap distributions - default = "orange".}

\item{int_colour}{Colour of the interval and summary point - default = "black".}
}
\value{
A ggplot object.
}
\description{
Plot hierarchical shift function percentile bootstrap samples generated with
\code{\link{hsf_pb}}. Bootstrap distributions of group trimmed mean
differences are plotted using tidybayes::geom_halfeyeh. The function returns
a ggplot object, which can be customised using the
\href{http://docs.ggplot2.org/current/}{ggplot2} package.
}
\examples{
plot_hsf_pb(out) # default plot
p <- plot_hsf_pb(out) # get ggplot object
plot_hsf_pb(out, gp_line_colour = "purple", gp_point_colour = "purple") # pink group shift function

}
\references{
Rousselet, G. A., & Wilcox, R. R. (2019, January 17). Reaction
  times and other skewed distributions: problems with the mean and the
  median. https://doi.org/10.31234/osf.io/3y54r
}
