% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_kde_rug_dec1}
\alias{plot_kde_rug_dec1}
\title{Plot 1 group: KDE + rug + deciles}
\usage{
plot_kde_rug_dec1(data = df, fill.colour = "grey30", fill.alpha = 0.3)
}
\description{
Plot kernel density estimate + rug plot + superimposed deciles
for 1 group stored in a data frame
}
