% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_kde_rug_dec2}
\alias{plot_kde_rug_dec2}
\title{Plot 2 groups: KDE + rug + deciles}
\usage{
plot_kde_rug_dec2(data = df)
}
\description{
Plot kernel density estimates + rug plots + deciles
for 2 groups stored in a data frame.
}

