% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function_plot.R
\name{plot_pbsf}
\alias{plot_pbsf}
\title{Plot shift function generated with shifthd_pbci or shiftdhd_pbci}
\usage{
plot_pbsf(df = data)
}
\description{
Plot shift function generated with shifthd_pbci or shiftdhd_pbci
}

