% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_quartile_bars}
\alias{plot_quartile_bars}
\title{Add bars marking the quartiles to plot}
\usage{
plot_quartile_bars(p, q_seq = c(0.25, 0.5, 0.75), col = "grey21",
  width = 0.5, q_size = 0.5, md_size = 1)
}
\description{
Add bars marking the quartiles to plot
}

