% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_scat2d}
\alias{plot_scat2d}
\title{Plot paired observations}
\usage{
plot_scat2d(df = df, xname = "condition1", yname = "condition2",
  min.x = NA, min.y = NA, max.x = NA, max.y = NA, axis.steps = 2,
  size_p = 5, stroke_p = 1, shape_p = 21, colour_p = "black",
  fill_p = "#ffb347", alpha_p = 0.5, linetype_q = "dashed", size_q = 1,
  alpha_q = 0.5, colour_q = "black")
}
\description{
Scatterplot of paired observations with reference line of no effect.
Quartiles of each condition are superimposed.
Input is a data frame with 3 columns: participant, condition1, condition 2.
Parameters _p control the scatterplot.
Parameters _q control the segments marking the quartiles.
}

