% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function_plot.R
\name{plot_sf}
\alias{plot_sf}
\title{Plot shift function generated with shifthd or shiftdhd}
\usage{
plot_sf(data = df, plot_theme = 1, xlabel = "Group 1 deciles",
  ylabel = "Decile differences", symb_col = NULL, symb_fill = NULL,
  symb_size = 5, symb_shape = 21, diffint_col = NULL,
  diffint_size = 0.5, dec_line_col = NULL, dec_line_alpha = 0.5,
  dec_line_size = 1.5, theme2_alpha = NULL)
}
\arguments{
\item{data}{A data frame generated by \code{shifthd} or \code{shiftdhd}.}

\item{plot_theme}{One of 3 themes - default is 1.}

\item{symb_col}{Contour colour of decile symbol.}

\item{symb_fill}{Fill colour(s) of decile symbol. Defaults are "white" for theme 1,
c("darkviolet","darkorange2") for theme 2, c("white", "grey") for theme 3.}

\item{symb_size}{Size of decile symbol - default = 5.}

\item{symb_shape}{Shape of the decile symbol - default = 21 (disc).}

\item{diffint_col}{Colour of the lines marking the decile difference intervals.}

\item{diffint_size}{Size of the lines marking the decile difference intervals
- default = .5.}

\item{dec_line_col}{Colour of the line joining the deciles.}

\item{dec_line_alpha}{Alpha of the line joining the deciles - default = .5.}

\item{dec_line_size}{Size of the line joining the deciles - default = 1.5.}

\item{theme2_alpha}{Alpha transparency for theme 2 - default c(0.4, 1)}

\item{xlab}{Label for the x axis.}

\item{ylab}{Label for the y axis.}
}
\description{
Plot a shift function using deciles, as generated by \code{shifthd} or
\code{shiftdhd}. The function returns a ggplot object, which can be
customised using the \code{ggplot2} package - see details
\href{http://docs.ggplot2.org/current/}{here}.
}
\details{
Several themes are available:
\itemize{
  \item Theme 1 uses a minimalist design with the same colour for all deciles.
  \item Theme 2 colour codes the sign of the differences.
  \item Theme 3 uses a greyscale gradient to code the deciles.
  }
}

