% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pxgta}
\alias{pxgta}
\title{Proportion of observations greater than a specified value}
\usage{
pxgta(x, a = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{a}{Value for comparison  - default = 0}
}
\value{
A list of 3 elements: \itemize{
\item \code{P(X>a)}
\item \code{P(X=a)}
\item \code{P(X<a)}
}
}
\description{
For a vector x, return the proportion of observations greater than a specified value a: P(X>a).
Also return the complementary proportions P(X<a) and P(X=a).
}
\section{Note}{

From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}
}

\seealso{
adapted from \code{\link{cid}}
}
