% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pxgty}
\alias{pxgty}
\title{Proportion of observations in x greater than observations in y}
\usage{
pxgty(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A list of 3 elements: \itemize{
\item \code{P(X>Y)}
\item \code{P(X=Y)}
\item \code{P(X<Y)}
}
}
\description{
For two vectors x and y, return the proportion of observations in x greater
than observations in y: P(X>Y). Also return the complementary proportions
P(X<Y) and P(X=Y). The proportions are determined by computing all pariwise
differences between vectors.
}
\section{Note}{

From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}
}

\references{
Cliff, N. (1996) Ordinal methods for behavioral data analysis. Erlbaum, Mahwah, N.J.
}
\seealso{
adapted from \code{\link{cid}}
}
