% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{q1469}
\alias{q1469}
\title{Estimates quantiles .1:.4 & .6:.9}
\usage{
q1469(x)
}
\description{
Estimate the deciles for the data in vector x using the Harrell-Davis
estimate of the qth quantile. Modified from \code{deciles} to return only
}

