% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function_calc.R
\name{shiftdhd}
\alias{shiftdhd}
\title{Shift function for two dependent groups (pbse methods)}
\usage{
shiftdhd(data = df, formula = obs ~ gr, nboot = 200)
}
\arguments{
\item{data}{A data frame in tidy format. Column 1 describes the two groups;
column 2 contains the values for each group. A properly formatted data
frame can be created using \code{link{mkt2}} or \code{link{mkt2d}}. Missing
values are not allowed.}

\item{formula}{A formula with format response variable ∼ predictor variable,
where ~ (tilde) means "is modeled as a function of".}

\item{nboot}{The number of bootstrap samples - default = 200.}
}
\value{
A data frame with one row per decile. The columns are: \itemize{
  \item Column 1 = deciles for group 1 \item Column 2 = deciles for group 2
  \item Column 3 = differences (column 1 - column 2) \item Column 4 = lower
  bounds of the confidence intervals \item Column 5 = upper bounds of the
  confidence intervals }
}
\description{
\code{shiftdhd} returns a shift function for two dependent groups.
\code{shiftdhd} \strong{works exclusively with deciles and with alpha =
0.05}, so that only 95\% confidence intervals can be computed To use other
quantiles and other alpha levels, see \code{\link{shiftdhd_pbci}}.
}
\details{
A shift function shows the difference between the quantiles of two groups as
a function of the quantiles of one group. For inferences, the function
returns a confidence interval for each quantile difference. The confidence
intervals are computed using a percentile bootstrap estimation of the
standard error of the difference between quantiles. The confidence intervals
are corrected for multiple comparisons so that the simultaneous probability
coverage is .95.

The deciles are estimated using the Harrell-Davis quantile estimator of the
qth quantile is used. The default number of bootstrap samples is nboot = 200.
The same bootstrap samples are used for each quantile and for each group.
}
\section{Note}{
 Modified from Rallfun-v32.txt - see
  \url{https://github.com/nicebread/WRS/}.
}
\examples{
out <- shiftdhd(df, formula = obs ~ gr) # use the default parameters
out <- shiftdhd(df, formula = obs ~ gr, nboot = 500) # specify the number of bootstrap samples

}
\references{
Harrell, F.E. & Davis, C.E. (1982) A new distribution-free
  quantile estimator. Biometrika, 69, 635-640.
  Wilcox, R.R. (2012) Introduction to robust estimation and hypothesis
  testing. Academic Press, San Diego, CA.
}
\seealso{
\code{\link{shifthd}} for independent groups.
}

