% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function_calc.R
\name{shifthd_pbci}
\alias{shifthd_pbci}
\title{Shift function for two independent groups (pbci method)}
\usage{
shifthd_pbci(data = df, formula = obs ~ gr, q = seq(0.1, 0.9, 0.1),
  nboot = 2000, alpha = 0.05, adj_ci = TRUE)
}
\arguments{
\item{q}{Quantiles to estimate - default = deciles 0.1:0.1:.9.
Other suggestions}

\item{nboot}{Number of bootstrap samples - default = 2000}

\item{alpha}{Expected long-run type I error rate - default = 0.05}

\item{x, y}{Vectors; no missing values are allowed}
}
\value{
A data frame with one row per decile.
The columns are: \itemize{
  \item Column 1 = quantiles for group 1
  \item Column 2 = quantiles for group 2
  \item Column 3 = differences (column 1 - column 2)
  \item Column 4 = lower bounds of the confidence intervals
  \item Column 5 = upper bounds of the confidence intervals
  \item Column 6 = p-values based on Hochberg's method
  \item Column 7 = critical p-values based on Hochberg's method
  }
}
\description{
Compute a shift function for two independent groups using the
Harrell-Davis quantile estimator in conjunction with a percentile
bootstrap approach.
Unlike \code{\link{shifthd}}: \itemize{
\item The confidence intervals are calculated using a percentile bootstrap of
the quantiles, instead of a percentile bootstrap of the standard error of
the difference of the quantiles.
\item The quantiles can be specified and are not limited to the deciles.
\item Tied values are allowed.
}
Examples of quantile sequences, from sparse to dense: \itemize{
\item \code{q = c(.25,.5,.75)}
\item \code{q = c(.1,.25,.5,.75,.9)}
\item \code{q = seq(.1, .9, .1)}
\item \code{q = seq(.05, .95, .05)}
}
}
\section{Note}{

Adaptation of Rand Wilcox's `qcomhd` & `pb2gen` R functions
(\url{http://dornsife.usc.edu/labs/rwilcox/software/}).
From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}.
}
\references{
Wilcox, R.R., Erceg-Hurn, D.M., Clark, F. & Carlson, M. (2014)
Comparing two independent groups via the lower and upper quantiles.
J Stat Comput Sim, 84, 1543-1551.
}
\seealso{
\code{hd}, \code{shifthd} for the pbse method for independent
  groups, \code{shiftdhd_pbci} for dependent groups
}

