% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{add_sf_lab}
\alias{add_sf_lab}
\title{Add difference labels to shift function}
\usage{
add_sf_lab(p, sf, labres = 2, link_col = c("darkviolet",
  "darkorange2"), link_alpha = c(0.4, 1), y_lab_nudge = 0,
  text_size = 5)
}
\arguments{
\item{p}{A list of ggplot objects generated by \code{\link{plot_sf}} or \code{\link{plot_pbsf}}.}

\item{sf}{A list of data frames generated by \code{\link{shifthd}},
\code{\link{shiftdhd}}, \code{\link{shifthd_pbci}} or
\code{\link{shiftdhd_pbci}}.}

\item{labres}{Number of decimales for the labels - default = 2.}

\item{link_col}{Label colours for negative and positive values.}

\item{link_alpha}{Alpha transparency of the labels - default = continuum between 0.4 and 1.}

\item{y_lab_nudge}{Amount by which to nudge the labels along the y axis.}

\item{text_size}{Text size - default = 5}
}
\description{
Add labels of quantile difference values to one or more shift function plots.
Used in the README.md file to illustrate the shift function. Assumes an odd
number of quantiles with the median in the middle.
}
