% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_asymmetry_calc.R
\name{asymhd}
\alias{asymhd}
\title{Difference asymmetry function for two independent groups}
\usage{
asymhd()
}
\description{
The difference asymmetry function provides perspective on the degree a distribution
is symmetric about zero, by quantifying the sum of q and 1-q quantiles of the
distribution of all pairwise differences between two independent groups x and y.
 If the distributions of x and y are identical, then the
 distribution of all pairwise differences is symmetric about zero.
If the distribution is symmetric the function should be approximately a horizontal line.
If in addition the median of the difference scores is zero, the horizontal line will
intersect the y-axis at zero.
Confidence intervals and p values are returned for each quantile sum.
The FWE is controlled via Hochberg's method, which is used to determine critical
p values based on the argument alpha.
}
\section{Note}{

This function combines Rand Wilcox's qwmwhd & cbmhd R functions,
from Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/},
http://dornsife.usc.edu/labs/rwilcox/software/.
}
\references{
Wilcox, R.R., Erceg-Hurn, D.M., Clark, F. & Carlson, M. (2014)
Comparing two independent groups via the lower and upper quantiles.
J Stat Comput Sim, 84, 1543-1551.
}
\seealso{
\code{\link{hd}}, \code{\link{diff_asym}}
}

