% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_dec_ci}
\alias{plot_dec_ci}
\title{Plot deciles and confidence intervals}
\usage{
plot_dec_ci(out = out, plotzero = TRUE, xtitle = "Differences",
  hjust = -0.05, vjust = 0.2, size_text = 6, colour_dec = "#009E73",
  fill_dec = "white", size_dec = 4, shape_dec = 21,
  colour_line = "#009E73", size_line = 1, linetype_line = 2,
  colour_zero = "black", size_zero = 0.5, linetype_zero = 1)
}
\description{
Plot deciles and their confidence intervals
using the output of `quantiles_pbci`
}

