% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_asymmetry_plot.R
\name{plot_diff_asym}
\alias{plot_diff_asym}
\title{Plot difference asymmetry functions}
\usage{
plot_diff_asym(data = df, symb_col = NULL, symb_fill = NULL,
  symb_size = 5, symb_shape = 21, diffint_col = NULL,
  diffint_size = 0.5, q_line_col = NULL, q_line_alpha = 0.5,
  q_line_size = 1.5)
}
\arguments{
\item{data}{A list of data frames generated by \code{asymhd} or \code{asymdhd}.}

\item{symb_col}{Contour colour of quantile symbol.}

\item{symb_fill}{Fill colour(s) of quantile symbol. Defaults are "white" for theme 1,
c("darkviolet","darkorange2") for theme 2, c("white", "grey") for theme 3.}

\item{symb_size}{Size of quantile symbol - default = 5.}

\item{symb_shape}{Shape of the quantile symbol - default = 21 (disc).}

\item{diffint_col}{Colour of the lines marking the quantile difference intervals.}

\item{diffint_size}{Size of the lines marking the quantile difference intervals
- default = .5.}

\item{q_line_col}{Colour of the line joining the quantiles.}

\item{q_line_alpha}{Alpha of the line joining the quantiles - default = .5.}

\item{q_line_size}{Size of the line joining the quantiles - default = 1.5.}
}
\value{
A list of ggplot objects.
}
\description{
Plot one or more difference asymmetry functions generated by \code{\link{asymhd}} or \code{\link{asymdhd}}. The function returns a list of ggplot objects, which can be
customised using the \code{ggplot2} package - see details
\href{http://docs.ggplot2.org/current/}{here}.
}
\examples{
# using asymhd
set.seed(21) # generate data
n <- 100 # sample size
df <- tibble(gr = factor(c(rep("group1",n),rep("group2",n),rep("group3",n))),
             obs= c(rnorm(n)+6, rnorm(n)+4, rnorm(n)*1.5+6)) # make tibble
out <- asymhd(df, obs ~ gr, doall = TRUE, nboot = 100) # compute all comparisons
plist <- plot_diff_asym(out)

# using asymdhd
set.seed(21) # generate data
n <- 100 # sample size per condition
C1 <- rnorm(n) # condition 1
C2 <- C1 + rnorm(n) + 2 # condition 2
# Data with 3 independent groups and 2 dependent conditions per group
library(tibble)
df <- tibble(gr = factor(c(rep("group1",n),rep("group2",n),rep("group3",n))),
  cond1 = c(C1, C1+rnorm(n), C1+rnorm(n)),
  cond2 = c(C2, C2 + 1, C2 + 3) ) # make tibble
library(dplyr)
df <- mutate(df, diff = cond1 - cond2)
out <- asymdhd(data = df, formula = diff ~ gr, nboot = 100, doall = TRUE)
plist <- plot_diff_asym(data = out)
plist

}
