% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_scat2}
\alias{plot_scat2}
\title{Scatterplots for 2 groups}
\usage{
plot_scat2(data = df, xlabel = NULL, ylabel = NULL, ...)
}
\description{
\code{plot_scat2} produces scatterplots for 2 marginal distributions.
The scatterplots are jittered using \code{\link[ggbeeswarm]{geom_quasirandom}}.
}
