% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_scat2d}
\alias{plot_scat2d}
\title{Plot paired observations}
\usage{
plot_scat2d(df = df, formula = cond2 ~ cond1, min.x = NA,
  min.y = NA, max.x = NA, max.y = NA, axis.steps = 2, size_p = 5,
  stroke_p = 1, shape_p = 21, colour_p = "black",
  fill_p = "#ffb347", alpha_p = 0.5, linetype_q = "dashed",
  size_q = 1, alpha_q = 0.5, colour_q = "black")
}
\arguments{
\item{df}{Data frame with paired observations in two columns.}

\item{formula}{A formula with format response variable ∼ predictor variable,
where ~ (tilde) means "is modeled as a function of".}

\item{min.x, min.y, max.x, max.y}{Specify axis limits - default square axes}

\item{axis.steps}{Steps between x and y tick marks - default = 2.}

\item{size_p}{Size parameter of the scatterplot - default = 5.}

\item{stroke_p}{Stroke parameter of the scatterplot - default = 1,}

\item{shape_p}{Shape parameter of the scatterplot - default = 21,}

\item{colour_p}{Colour parameter of the scatterplot - default "black".}

\item{fill_p}{Fill parameter of the scatterplot - default = "#ffb347",}

\item{alpha_p}{Alpha parameter of the scatterplot - default = .5,}

\item{linetype_q}{Linetype of the segments marking the quartiles - default = "dashed",}

\item{size_q}{Size of the segments marking the quartiles - default = 1,}

\item{alpha_q}{Alpha of the segments marking the quartiles - default = .5,}

\item{colour_q}{Colour of the segments marking the quartiles - default = "black"}

\item{colour_p}{Colour parameter of the scatterplot - default = "black",}
}
\description{
Scatterplot of paired observations with reference line of no effect.
Quartiles of each condition are superimposed. Quartiles are estimated using the
Harrell-Davis estimator.
}
\examples{
df <- tibble(cond1 = rnorm(50), cond2 = cond1 + rnorm(50))
plot_scat2d(df, formula = cond2 ~ cond1) # basic call
plot_scat2d(df, formula = cond2 ~ cond1, size_q=3) # specify size of quartile segments
plot_scat2d(df, formula = cond2 ~ cond1, size_q=c(1,2,1)) # use thicker line for median
plot_scat2d(df, formula = cond2 ~ cond1, linetype_q = "longdash") # specify linetype - default = dashed
}
\seealso{
\code{\link{hd}}
}
