% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_scat2d}
\alias{plot_scat2d}
\title{Plot paired observations}
\usage{
plot_scat2d(df = df, xname = "condition1", yname = "condition2",
  min.x = NA, min.y = NA, max.x = NA, max.y = NA, axis.steps = 2,
  psize = 5, pstroke = 1, pshape = 21, pcolour = "black",
  pfill = "#ffb347", palpha = 0.5)
}
\description{
Scatterplot of paired observations with reference line of no effect.
Quartiles of each condition are superimposed.
Input is a data frame with 3 columns: participant, condition1, condition 2
}

