% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{plot_scat2d}
\alias{plot_scat2d}
\title{Plot paired observations}
\usage{
plot_scat2d(df = df, xcol = "condition1", ycol = "condition2",
  min.x = NA, min.y = NA, max.x = NA, max.y = NA, axis.steps = 2,
  size_p = 5, stroke_p = 1, shape_p = 21, colour_p = "black",
  fill_p = "#ffb347", alpha_p = 0.5, linetype_q = "dashed", size_q = 1,
  alpha_q = 0.5, colour_q = "black")
}
\arguments{
\item{df}{Data frame with paired observations in two columns.}

\item{xcol, ycol}{Names of columns of paired observations.}

\item{colour_p}{Colour parameter of the scatterplot.}

\item{fill_p}{Fill parameter of the scatterplot and so on.}

\item{linetype_q}{Linetype of the segments marking the quartiles and so on.}

\item{colour_q}{Colour of the segments marking the quartiles.}
}
\description{
Scatterplot of paired observations with reference line of no effect.
Quartiles of each condition are superimposed. Quartiles are estimated using the
Harrell-Davis estimator.
}
\examples{
plot_scat2d(df, xcol='c1', ycol='c2') # basic call
plot_scat2d(df, xcol='c1', ycol='c2', size_q=1) # specify size of quartile segments
plot_scat2d(df, xcol='c1', ycol='c2', size_q=c(1,2,1)) # use thicker line for median
plot_scat2d(df, xcol='c1', ycol='c2', linetype_q = "longdash") # specify linetype - default = dashed
}
\seealso{
\code{\link{hd}}
}
