% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function_plot.R
\name{plot_sf}
\alias{plot_sf}
\title{Plot percentile bootstrap shift function}
\usage{
plot_sf(data = df, plot_theme = 1, symb_col = NULL, symb_fill = NULL,
  symb_size = 5, symb_shape = 21, diffint_col = NULL,
  diffint_size = 0.5, q_line_col = NULL, q_line_alpha = 0.5,
  q_line_size = 1.5, theme2_alpha = NULL)
}
\arguments{
\item{data}{A list of data frames generated by \code{shifthd_pbci} or \code{shiftdhd_pbci}.}

\item{plot_theme}{One of 3 themes - default is 1.}

\item{symb_col}{Contour colour of quantile symbol.}

\item{symb_fill}{Fill colour(s) of quantile symbol. Defaults are "white" for theme 1,
c("darkviolet","darkorange2") for theme 2, c("white", "grey") for theme 3.}

\item{symb_size}{Size of quantile symbol - default = 5.}

\item{symb_shape}{Shape of the quantile symbol - default = 21 (disc).}

\item{diffint_col}{Colour of the lines marking the quantile difference intervals.}

\item{diffint_size}{Size of the lines marking the quantile difference intervals
- default = .5.}

\item{q_line_col}{Colour of the line joining the quantiles.}

\item{q_line_alpha}{Alpha of the line joining the quantiles - default = .5.}

\item{q_line_size}{Size of the line joining the quantiles - default = 1.5.}

\item{theme2_alpha}{Alpha transparency for theme 2 - default c(0.4, 1)}
}
\description{
Plot one or more shift functions generated with \code{\link{shifthd}},
\code{link{shiftdhd}}, \code{\link{shifthd_pbci}} or
\code{link{shiftdhd_pbci}}. Assumes the median was estimated and is the
middle value. The function returns a list of ggplot objects, which can be
customised using the \href{http://docs.ggplot2.org/current/}{ggplot2}
package.
}
\details{
Several themes are available:
\itemize{
  \item Theme 1 uses a minimalist design with the same colour for all quantiles.
  \item Theme 2 colour codes the sign of the differences.
  \item Theme 3 uses a greyscale gradient to code the quantiles.
  }
}
\examples{
plot_sf(out) # default plot
plot_sf(out, plot_theme = 2) # specify theme
plist <- plot_sf(out) # output list of ggplot objects

# The plots can then be combined using \\href{https://cran.r-project.org/web/packages/gridExtra/index.html}{gridExtra} or \\href{https://cran.r-project.org/web/packages/cowplot/index.html}{cowplot} - for instance:
library(gridExtra)
do.call("grid.arrange", c(plist, ncol=2))
# To extract one object and for instance change a label:
p <- plist[[1]]
p + labs(y = "Difference")

}
