% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_function_calc.R
\name{shiftdhd_pbci}
\alias{shiftdhd_pbci}
\title{Shift function for two depend groups (pbci methods)}
\usage{
shiftdhd_pbci()
}
\description{
This version of the shift function
Computes a shift function for two dependent groups by comparing
the quantiles of the marginal distributions using the
Harrell-Davis quantile estimator in conjunction with a percentile
bootstrap approach.
}
\details{
INPUTS:
- x & y are vectors of the same length without missing values
- q = quantiles to estimate - default = 0.1:0.1:.9 (deciles)
- nboot = number of bootstrap samples - default = 2000
- alpha = expected long-run type I error rate - default = 0.05
- plotit = 1 to get a figure; 0 otherwise by default

OUTPUTS:
 - xd & yd = vectors of quantiles
- delta = vector of differences between quantiles (x-y)
- deltaCI = matrix quantiles x low/high bounds of the confidence
intervals of the quantile differences

Unlike shiftdhd:
- the confidence intervals are not corrected for multiple comparisons
(the R version provides corrected critical p values - here i've decided
not to provide p values at all - the goal is to understand how
distributions differ, not to make binary decisions)
- the confidence intervals are calculated using a percentile bootstrap of
  the quantiles, instead of a percentile bootstrap of the standard error of the quantiles
- the quantiles to compare can be specified and are not limited to the
  deciles
- Tied values are allowed

Unlike Rand Wilcox's Dqcomhd R function, no p value is returned.
 Extensive experience suggests humans cannot be trusted with p values.
}
\section{Note}{

Adaptation of Rand Wilcox's Dqcomhd, bootdpci & rmmcppb R functions,
http://dornsife.usc.edu/labs/rwilcox/software/
From Rallfun-v32.txt - see \url{https://github.com/nicebread/WRS/}
}
\references{
Wilcox, R.R. & Erceg-Hurn, D.M. (2012)
Comparing two dependent groups via quantiles.
J Appl Stat, 39, 2655-2664.
}
\seealso{
HD, SHIFTDHD, SHIFTHD_PBCI
}

